backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.51.0
google-generativeai==0.8.3
python-dotenv==1.0.0
requests==2.31.0
pymongo==4.6.3
sentence-transformers==2.7.0
numpy==1.24.3
pandas==2.0.3
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
PyMuPDF==1.23.26
Pillow==10.1.0
google-cloud-vision==3.4.5
